% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNF.R
\name{CNF}
\alias{CNF}
\title{Compute the marginal loglikelihood of the multivariate count outcome.}
\usage{
CNF(Ct, g, method)
}
\arguments{
\item{Ct}{A vector of categorical count.}

\item{g}{A function of expected values for each category of the multivariate count outcome.}

\item{method}{An argument, either "dirmult" for the Dirichlet-multinomial mixed model and "multl" for the multinomial logistics mixed model.}
}
\value{
The loglikelihood values for the marginal distribution of the multivariate count outcome.
}
\description{
Compute the marginal loglikelihood of the multivariate count outcome.
}
