% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerate.R
\name{DataGenerate}
\alias{DataGenerate}
\title{Generate observation for joint model between multivariate count outcome and continuous outcome
The marginal distribution for the multivariate count outcome could be either Dirichlet - multinomial mixed model or multinomial logistic mixed model}
\usage{
DataGenerate(N, method, var.level)
}
\arguments{
\item{N}{A number of subjects.}

\item{method}{An argument defining the marginal distribution of the multivariate count outcome, either "dirmult" or "multl".}

\item{var.level}{An argument of either "high" or "low" as defined in the simulation setting}
}
\value{
A dataset in wideformat consisting the observations of the multivariate count outcome for two time points, the covariate and the continuous outcome.
}
\description{
Generate observation for joint model between multivariate count outcome and continuous outcome
The marginal distribution for the multivariate count outcome could be either Dirichlet - multinomial mixed model or multinomial logistic mixed model
}
\examples{
DataGenerate(100,method = "dirmult",var.level = "high")
}
