% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.R
\name{F}
\alias{F}
\title{A loglikelihood value of a subject given a random effect \code{z}}
\usage{
F(z, b, theta, Sigma, ev, data, method)
}
\arguments{
\item{z}{A vector of random effect .}

\item{b}{A vector of fixed effect parameters.}

\item{theta}{an overdispersion parameter for the Dirichlet - multinomial mixed model.}

\item{Sigma}{The matrix of covariance of the random effect.}

\item{ev}{the standard deviation for residuals in continous outcome.}

\item{data}{a vector of observation for a subject.}

\item{method}{An argument of either "dirmult" for Dirichlet-multinomial mixed model or "multl" for multinomial logistics mixed model.}
}
\description{
A loglikelihood value of a subject given a random effect \code{z}
}
