% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixEf.R, R/integralComputation.R
\name{FixEf}
\alias{FixEf}
\title{Linear combination of the fixed effect for each categories in multivariate count outcome.}
\usage{
FixEf(x.vec, b, Des, method)

FixEf(x.vec, b, Des, method)
}
\arguments{
\item{x.vec}{a value of binary covariate.}

\item{b}{vector of regression coefficients.}

\item{Des}{a design matrix}

\item{method}{either "dirmult" for Dirichlet - multinomial or "multl" for multinomial logistics mixed model.}
}
\value{
a vector representing the result of multiplication between \code{Des} and \code{b} according to value of \code{x.vec} and the chosen method.

a vector representing the result of multiplication between \code{Des} and \code{b} according to value of \code{x.vec} and the chosen method.
}
\description{
Linear combination of the fixed effect for each categories in multivariate count outcome.

Linear combination of the fixed effect for each categories in multivariate count outcome.
}
\examples{
Des <- as.matrix(cbind(rep(1,6),rep(0:1,3),c(0,0,1,1,0,0), c(rep(0,4),rep(1,2)),c(0,0,0,1,0,0),c(rep(0,5),1)))
beta <- c(0,0.5,-3.5,-1.3,0.8,-0.15,-2,1)
FixEf(x.vec = 1,b = beta, Des = Des,method = "dirmult")
Des <- as.matrix(cbind(rep(1,6),rep(0:1,3),c(0,0,1,1,0,0), c(rep(0,4),rep(1,2)),c(0,0,0,1,0,0),c(rep(0,5),1)))
beta <- c(0,0.5,-3.5,-1.3,0.8,-0.15,-2,1)
FixEf(x.vec = 1,b = beta, Des = Des,method = "dirmult")
}
