% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgauss_hermite.R
\name{mgauss_hermite}
\alias{mgauss_hermite}
\title{Nodes and weights for multivariate Gauss - Hermite quadratures}
\usage{
mgauss_hermite(n, mu, sigma)
}
\arguments{
\item{n}{the number of quadratures}

\item{mu}{mean vector.}

\item{sigma}{covariance matrix}
}
\value{
a set of points and nodes of multivariate Gauss-Hermite quadratures corresponding to the mean vector \code{mu} and \code{sigma}
}
\description{
Nodes and weights for multivariate Gauss - Hermite quadratures
}
\examples{
Sigma <- cbind(c(1.5,-0.6),c(-0.6,4))
mgauss_hermite(n = 3, mu = c(1.5,0), sigma = Sigma)
}
